refactor: clean up quote plugin
